% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africa_data_all.R
\docType{data}
\name{africa_data_2020}
\alias{africa_data_2020}
\title{2020 data on 54 African countries and 4 dependencies/territories}
\format{
A tibble with 58 rows and 11 variables
\describe{
\item{country}{chr Name of country}
\item{pop}{dbl Country's population (2020)}
\item{pop.yearly.change}{dbl Country's yearly population change}
\item{pop.net.change}{dbl Country's net population change}
\item{density}{dbl Country's population density (P/Km²)}
\item{area}{dbl Total land area (Km²)}
\item{migrants}{dbl Net migrant population}
\item{fertility.rate}{dbl Fertility rate}
\item{med.age}{dbl Median age of population}
\item{urban.pop}{dbl Percent of population in urban areas}
\item{world.share}{dbl Country's world share of population}
}

@source The most up-to-date data for these variables can be found at \url{https://www.worldometers.info/population/countries-in-africa-by-population/}
}
\usage{
africa_data_2020
}
\description{
Countries and dependencies are classified by population (pop), yearly change, net change, density, area, migrants, fertility rate, median age (med.age), urban poplation, and world share of population.
}
\keyword{datasets}
