% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africa_subregion.R
\docType{data}
\name{africa_subregion}
\alias{africa_subregion}
\title{Data on 54 African countries and their subregions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54 rows and 4 columns.
}
\usage{
africa_subregion
}
\description{
Countries are classified by population and subregion.
}
\details{
A tibble with 54 rows and 4 variables

\describe{
\item{\verb{#}}{dbl Rank order by population size}
\item{Country}{chr Name of country}
\item{Population}{dbl Overall total population (2020)}
\item{Subregion}{char Location of country in African subregions}
}
}
\keyword{datasets}
