% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmedianincome.R
\docType{data}
\name{blackmedianincome}
\alias{blackmedianincome}
\title{Sample Census data B19013B_001 black median income 2020}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 52 rows and 5 columns.
}
\usage{
blackmedianincome
}
\description{
Median income estimate is classified by state and includes margin of error (moe).
}
\details{
A tibble with 52 rows and 5 variables

\describe{
\item{GEOID}{chr State geographic identification number}
\item{NAME}{chr State name}
\item{variable}{chr Variable name}
\item{Population}{dbl Estimate (2020)}
\item{moe}{dbl Margin of error}
}
}
\keyword{datasets}
