% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmedianincome_quintile_calc.R
\docType{data}
\name{blackmedianincome_calc}
\alias{blackmedianincome_calc}
\title{blackmedianincome state quintiles}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 52 rows and 6 columns.
}
\usage{
blackmedianincome_calc
}
\arguments{
\item{data}{Dataset; must have our predefined column names: GEOID to moe}
}
\value{
Vector of quntile
}
\description{
Calculating state quntiles for black median income variable
}
\examples{
blackmedianincome$estimate_quintile <- blackmedianincome_calc()
}
\keyword{datasets}
