% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continents_pop_2010.R
\docType{data}
\name{continents_pop_2010}
\alias{continents_pop_2010}
\title{2010 continental populations}
\format{
A tibble with 7 rows and 5 variables
\describe{
\item{continent}{chr Name of continent}
\item{area.sq.km}{dbl Land area in square km}
\item{area.sq.mi}{dbl Land area in square mi}
\item{pop}{dbl 2010 population}
\item{pop.density}{dbl NAs}
}

@source United Nations population estimates
}
\usage{
continents_pop_2010
}
\description{
Continental populations in 2010
}
\keyword{datasets}
