% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continents_pop_2021.R
\docType{data}
\name{continents_pop_2021}
\alias{continents_pop_2021}
\title{2021 continental populations}
\format{
A tibble with 7 rows and 4 variables

\describe{
\item{continent}{chr Name of continent}
\item{pop}{dbl 2010 population}
\item{percent.world.pop}{dbl Percent of global population}
\item{growth.rate}{dbl Per annum growth rate}
}

@source United Nations population estimates
}
\usage{
continents_pop_2021
}
\description{
Continental populations in 2021
}
\keyword{datasets}
