% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continents_pop_2023.R
\docType{data}
\name{continents_pop_2023}
\alias{continents_pop_2023}
\title{2023 continental populations}
\format{
A tibble with 6 rows and 5 variables

\describe{
\item{continent}{chr Name of continent}
\item{pop}{dbl 2023 population}
\item{area.sq.km}{dbl Land area square km}
\item{pop.density.km2}{dbl Population density in square km}
\item{world.share}{dbl Share of global population}
}

@source United Nations population estimates
}
\usage{
continents_pop_2023
}
\description{
Continental populations in 2023
}
\keyword{datasets}
