% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critstats-package.R
\docType{package}
\name{critstats-package}
\alias{critstats}
\alias{critstats-package}
\title{critstats: Data pacakge for exercises in critical statistical analysis}
\description{
This package provides an assorment of data sets to accompany the ICS text in preparation by Nathan Alexander, PhD.
}
\author{
\strong{Maintainer}: Nathan Alexander \email{nna2106@columbia.edu} (\href{https://orcid.org/0000-0002-2029-7329}{ORCID})

}
\keyword{internal}
