% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presidential_elections.R
\docType{data}
\name{presidential_elections}
\alias{presidential_elections}
\title{Data on US presidential elections by democratic vote since 1932}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1097 rows and 4 columns.
}
\usage{
presidential_elections
}
\description{
States are classified by democratic vote, year, and geography
}
\details{
A tibble with 1,097 rows and 4 variables

\describe{
\item{state}{chr State name}
\item{demVote}{dbl Percent of democratic party vote}
\item{year}{dbl Election year}
\item{south}{lgl T/F indicator of state location in US south}
}
}
\keyword{datasets}
