% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxpoll.R
\docType{data}
\name{taxpoll}
\alias{taxpoll}
\title{US Poll on who it is better to raise taxes on}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 691 rows and 2 columns.
}
\usage{
taxpoll
}
\description{
Public Policy Polling, Americans on College Degrees, Classic Literature, the Seasons, and More, data collected Feb 20-22, 2015.
}
\details{
A tibble with 691 rows and 2 variables

\describe{
\item{party}{chr Political party affiliation}
\item{taxes}{chr Response for who to raise taxes on}
}
}
\keyword{datasets}
