% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_size.R
\docType{data}
\name{true_size}
\alias{true_size}
\title{Data on 18 countries and their land area in relation to the African continent}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 4 columns.
}
\usage{
true_size
}
\description{
Countries are detailed by percent of African continent land area.
}
\details{
A tibble with 18 rows and 4 variables

\describe{
\item{Country}{chr Name of country}
\item{percent.africa}{dbl Land proportion of African continent}
\item{area.sq.km}{dbl Land area of country in square km}
\item{area.sq.mi}{dbl Land area of country in square mi}
}
}
\keyword{datasets}
