% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_presidents.R
\docType{data}
\name{us_presidents}
\alias{us_presidents}
\title{Data on US Presidents through 2023}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 70 rows and 5 columns.
}
\usage{
us_presidents
}
\description{
Data contains president, president political party, start data, end data, and vice president by term
}
\details{
A tibble with 70 rows and 5 variables

\describe{
\item{potus}{chr President}
\item{party}{dbl President political party}
\item{start}{dbl Start data of presidential term}
\item{end}{dbl End date of presidential term}
\item{vpotus}{dbl Vice President}
}
}
\keyword{datasets}
