% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_pop.R
\docType{data}
\name{world_pop}
\alias{world_pop}
\title{World Bank global population data by country, 1960-2020}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 216 rows and 62 columns.
}
\usage{
world_pop
}
\description{
Data on the population of countries by year from 1960 to 2020.
}
\details{
A tibble with 216 rows and 62 variables

\describe{
\item{country}{chr Name of country}
\item{year_YYYY}{dbl population in YYYY (labels to be updated)}
}
}
\keyword{datasets}
